+++ CHANGELOG.md
@@ -25,6 +25,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Added functionality to `@cumulus/db/src/search` to support term queries
   - Updated `BaseSearch` and `GranuleSearch` classes to support term queries for granules
   - Updated granules List endpoint to search postgres
+- **CUMULUS-3696**
+  - Added functionality to `@cumulus/db/src/search` to support terms, `not` and `exists` queries
 
 ### Migration Notes